Fix wrong name in subrepl issue
Template quotes need a name defined in the reader to work. The LisspREPL always assumed '__main__', but could have been embedded in some other module, so this assumption was wrong. Now assumes '__name__' (all modules have this, or they're broken), and falls back to '__main__' in case it was started with an empty namespace.