Change | short name from BAR to VERT
While Clojure's munge calls it that, there are a lot of characters with "BAR" in the Unicode name and | isn't one of them. It's called "VERTICAL LINE". If "LINE" isn't a good name, "BAR" isn't either.

"bar" is also a common metasyntactic variable/placeholder name, which should be avoided in "real-world" usage.

&vert; is the HTML entity, so there's good precedent.