Add macroexpand1 error check for missing ns
This seems like it would be a common mistake for those coming from other Lisps, and the reason why wasn't obvious enough from the traceback.

Nicer behavior for REPL debugging usage (when you wouldn't have a macro context) might be to use the current globals, but macroexpand and friends would have to walk up the stack to find it, which seems error-prone, especially for the recursive macroexpand_all.