Eliminate the DROP mechanism
The implementation of _# no longer requires it. Originally meant for side effects, it was of dubious value considering that reader macros should ideally be idempotent. (The SoftSyntaxError mechanism requires re-reads.) It's still possible for an arity-2 tag to simply return only one of its args for a similar effect. A major planned use for side-effecting reads was for implementing multiary tags, but tags have that feature built in now.