Require escaped or alnum kwarg token penultimate character
This rule is a little inelegant, but does the right thing in most cases. Kwargs will usually be Python identifiers. `foo!=bar` is a symbol. `foo\!=bar` is a Kwarg. Symbols like `=>` and `<=` are allowed now. `foo=>bar` is a Kwarg, but `foo[=>]bar`, `foo\=>bar`, and `foo<=bar` are symbols. Starargs are now a separate token, but still yield a Kwarg object.