Implement progn optimization
A zero-argument call of a zero-parameter lambda form can be optimized to just its body.

While it would make the compiled output look a little cleaner, it's not clear that unpacking nested progns is any more efficient, and it would be a lot harder to implement. Nesting saves a little stack memory at the cost of a few more instructions, so could actually be useful when locals are large.