Change compiled fromlist to '*'
CPython seems to ignore this by default, but import hooks might use it?

Unlike '?', a '*' is probably always valid, because this is close to the `('*',)` that a `from foo import *` statement would produce. No other identifier is guaranteed to exist for every module. Strings are sequences like tuples, so `iter(fromlist)` and `fromlist[0]` work the same way with just '*'.