Compute __package__ from name in new_env()
This doesn't work correctly if the module itself is a package. But the __init__ file is not supposed to be compiled, because it's supposed to have the transpile call in the first place. This is not a serious limitation, because `__init__.py` can do an `import *` from a Lissp module imported after the transpile.