Fix kwarg token escape handling
Now it munges and processes escapes like a symbol first, which is more like a normal tag.

Previously, the tag made a Kwarg object with the verbatim token, sans trailing `=`, with the munging and escape processing only applied if used as a keyword argument for a tag.

 However, the Kwarg() type has uses in metaprograms, e.g., `my#`, with a similar use demonstrated in the macro tutorial. The difference in handling a real tag argument and a kwarg token inside the scope when it needs munging or escape processing is unintended and therefore a bug.

The correct fix is to process earlier in the reader, rather than require a extra steps to munge and process escapes for all future tags using this feature. Stararg tokens were already tokenized separately, and don't munge.