Fix REPL fragment continuation
For some fragments, the REPL previously incorrectly asked for more lines with no clear way to end the input. Some fragments triggering the issue, like |if 1:|, should be errors, while others should not, like |if 1:1|.

The REPL now attempts to add an extra newline to the compilation if Lissp didn't think it needed more lines of input, but the underlying Python interpreter thinks the compiled input isn't done. If that still doesn't work, the REPL will compile the Python directly to generate an exception, which will hopefully have a useful error message.