[Untested] Replace ors in destruct-> examples with ||
Applied to one argument, the empty invocation is an identity transform, compiling as operation-ordering parentheses. This is arguably a bit simpler than alternatives like ors, which requires a macroexpansion (and that the macro be in scope), although that only adds a comment, or worse alternatives, like doto, which compile to a lambda.